Make playback_info a dataclass and move dataclasses to their own folder
